Separate app from lib
